// web/src/react-shared/src/models/transaction.tsx
export enum TransactionType {
  ApproveShield = 'Approve Shield',
  Shield = 'Shield',
  Unshield = 'Unshield',
  Send = 'Send',
  Swap = 'Swap',
  ApproveSpender = 'ApproveSpender',
  Mint = 'Mint',
  Cancel = 'Cancel',
  FarmDeposit = 'Farm Deposit',
  FarmRedeem = 'Farm Redeem',
  AddLiquidity = 'Add Liquidity',
  RemoveLiquidity = 'Remove Liquidity'
}